added 3wayIndependentTrained to ways
# Intro
We originally trained a single classifier on 4 item categories - t1/t2/c1/c2. Or some variation. What we want to do is train two classifiers: t1/c1/c2 and t2/c1/c2. We store as t1_prob the Trained1 scores, as t2_prob the Trained2 scores, and as c1_prob and c2_prob the mean of the controlA and controlB scores, respectively. The result is that our t1 and t2 scores are a bit less statistically dependent upon one another (though I believe the presence of c1 and c2 in our classifiers mean there is still some dependency?).

# Hanging question: 
What to do about controls? Is it right to take the mean? Or should we store results of both clfs? Or discard controls completely? For now, I'll take mean. 

# The Canonical Way (4way)
	1. Normalize DF and RF (DF is drawing features, and RF is recognition run features)
	2. Define train/test split (training data comes from recognition runs, test data from drawing runs) and train classifier.
	3. Add prediction probabilities to metadata matrix (sorting so trained are first and control is last)
	4. Assign to out either the logprobs or the probs and and with out populate the DM output.
	5. Populate t1_prob t2_prob, etc and bed_prob, bench_prob etc in DM.

# What's different for 3way?
Loop through control objects. 
Remove from RM and RF data associated with the excluded control. 
Otherwise, just produce two clfprob sets, one for each control object loop. Take the mean. 
Another Difference: 3way doesn't assign anything to bed/bench/chair/table. Is this on purpose? It looks like there's no clean mapping to bed/bench/chair/table that way. Do we ever use bed/bench/chair/table later?

# What's different for 2way?
Subsets recognition data matrices to only include the trained classes.
Otherwise identical. Also doesn't assign to bed/bench/chair/table.

# How about 2wayDraw?
This one's really different from the others, I won't bother to make sense of it.

# Now how about 3wayIndependentTrained?
Can be conceptualized as congruent with 3way except with two big exceptions:

Like the 3way, we want to generate two clfs and their sets of probabilities. 
But whereas for the 3way, the difference between the underlying classifiers is which control object gets excluded (something unique to the subject/roi),  the difference between the underlying clfs in our 3wayIndependentTrained case is which *trained* object gets excluded.

And then there's a difference in how values are finally assigned to t1_prob t2_prob, etc.
Whereas for the 3way, the stored values are ultimately the mean of corresponding probabilities output from the two classifiers, in our case: (assuming out[0] is the clf including just t1, c1, & c2 - in that order; and out[1] is the clf including just t2, c1, c2 - in that order)
	- DM[t1_prob] is solely out[0][:,0]
	- DM[t2_prob] is solely out[1][:,0]
	- DM[c1_prob] is (out[0][:,1] + out[1][:,1])/2.0
	- DM[c2_prob] is (out[0][:,2] + out[1][:,2])/2.0

In other words, trained object probabilities are pulled from the classifier they're individually present in, while control object probabilities are produced by both classifiers so we take the mean of the two results.